// src/apis/types.service.ts
export interface ISearchCommon {
  label: string;
  value: string | number;
  keySearch: string;
}
// ...
export interface ISearchInformation {
  title?: string;
  type?: string;
  header_images?: IHeaderImageInfo[];
  description?: string;
  local_map?: ILocationMapInfo;
  weather?: string;
  population?: string;
  area_code?: string;
  climate?: string;
  postal_code?: string;
  region?: string;
  founded?: string;
  buttons?: ILocationButtons[];
}

export interface ILocationImage {
  link: string;
  source: string;
  thumbnail: string;
  original: string;
  title: string;
  source_name: string;
}

export interface IHeaderImageInfo {
  image: string;
  source: string;
}

export interface ILocationMapInfo {
  image: string;
  link: string;
}

export interface IResultInfo {
  title: string;
  link: string;
  thumbnail: string;
  snippet: string;
}

export interface ILocationButtons {
  text?: string;
  title?: string;
  link?: string;
  displayed_link?: string;
  snippet?: string;
  snippet_highlighted_words?: string[];
}
// ...
export interface IBusinessInfo {
  CUSTOMERID: string;
  COMPANYNAME: string;
  TAXCODE?: string;
  TEL?: string;
  FAX?: string;
  ADDRESS?: string;
  BANKNAME?: string;
  BANKACCOUNT?: string;
  REPRESENTATIVE?: string;
  REGION_ID?: number;
}

export interface IUnit extends ISearchCommon {
  UNITID: string;
  SHORTNAME?: string;
  NOSIGNSHORTNAME?: string;
  DESCRIPTION?: string;
  ACTIVE: string;
  USERCREATE?: string;
  CREATEDATE?: Date;
  REGION_ID: string;
  LOCATIONS?: ILocation[];
  UNITWITHNAME?: string;
}
// ...
export interface ILocation extends ISearchCommon {
  LOCATIONID?: string;
  LOCATIONNAME?: string;
  CREATEDDATE?: Date;
  USERCREATE?: string;
  ACTIVE: number;
  LOCATIONADDRESS?: string;
  PROVINCEID: string;
  TELL?: string;
  FAX?: string;
  EMAIL?: string;
  ID?: string;
  UNITID?: string;
  UNITNAME?: string;
}
// ...
export interface ICustomerDetail {
  CUSTOMERID?: string;
  CUSTNAME?: string;
  NOSIGNCUSTNAME?: string;
  ADDRESS: string;
  TEL?: string;
  IDCARD?: string;
  IDCARDDATE?: Date;
  BANKNAME?: string;
  BANKACCOUNT?: string;
  ISBUSINESS?: string;
  BUSINESSINFOS?: IBusinessInfo;
  ISGENERAL?: string;
  USERCREATE?: string;
  CREATEDATE?: Date;
  ACTIVE?: string;
  EMAILADDRESS?: string;
  NOTRECEIVEINVOICE?: string;
  REGION_ID?: number;
  UNITS?: IUnit[];
  PROVINCEID? : string;
  PROVINCENAME? : string;

  ACTION_TYPE: string;
  APPROVED_BY?: string; 
  APPROVED_DATE?: string; 
  ASSET?: string; 
  BANK_ACCOUNT?: string; 
  BRANCHID?: string; 
  BUSINESS_TYPE?: string; 
  CAPACITY?: string; 
  CHECKIN_ID?: string; 
  COMMUNE_ID?: string; 
  CREATED_BY?: string; 
  CREATED_DATE?: Date; 
  CUSTIDCARD: string;
  CUSTIDCARDDATE: Date;
  CUSTOMER_BUSINESS?: string; 
  DISTRICT_ID?: string; 
  EDOC_ID?: number; 
  EMAIL: string; 
  IS_APPROVED?: string; 
  IS_INSIGNING?: string; 
  NAME: string; 
  OTHERS?: string;
  OWNER_ADDRESS?: string;
  OWNER_NAME?: string;
  OWNER_PHONE?: string; 
  PHONE_NUMBER: string; 
  PROVINCE_ID?: string; 
  SIGN_STATUS?: number;
  STREET_ADDRESS?: string; 
  TAX_CODE?: string; 
  TMP_CODE?: number; 
  WARRANTY?: string;
  ListBranch?: ListBranch[];
  BRANCHIDLIST?: string;
}

export interface ListBranch {
  id?: string;
  name?:string;
  isSelected?:string;
}