// src/redux/master/master.types.ts
export interface IMasterState {
  userOfficeS: IMasterResponseCommon[];
  userDepartmentS: IMasterResponseCommon[];
  customers: ICustomerInfo[];
  customersTmp: ICustomerTmpInfo[];
  customerModel : ICustomerDetail;
  farmList : IFarmItem[];
  locations: ILocation[];
  saleLocations: ILocation[];
  productsUnit: IMasterResponseCommon[];
  productsDropdownData: DropdownItemType[];
  customerDropdownData: DropdownItemType[];
  locationsDropdownData: DropdownItemType[];
  prices: IPrice[];
  customerBalances: ICustomerBalance[];
  discountTypeDropdownData: DropdownItemType[];
  unitsOfCustomer: IUnit[];
  unitsCustomerDropdownData: DropdownItemType[];
  locationsOfUser: ILocation[];
  userMenus: IMasterMenu[];
  userMenuIds: number[];
  unitsSale: IUnit[];
  isSubmitSuccess: boolean;
  products: IProductStockDTO[];
  currentScreen: string;
  checkingID: number;
  checkoutID: number;
  checkingData: ICheckingHistoryItem[];
  checkingDataDetail: ICheckingHistoryDetail[];
  checkingImg: string;
  isCheckOut: boolean,
  imageID: string;
  QRcode: string;
  FarmID: string;
  lastCheckinResponse: ILastCheckinResponse[];
  AllReport: IAllReport[];
  cityLatLong: ICityLatLong[];
  Information: ISearchInformation;
  InformationButtons: ILocationButtons[];
  GoogleResult: IGoogleResult;
  TopSights: ITopSights;
  ImgurResult: IImgurResult;
  GoogleImgResult: IGoogleImgResult;
  ChatGPTResponse: IChatGPTResponse[];
}
// ...
export interface IGetImgUrlSuccess {
  type: Types.MASTER_GET_IMG_URL_SUCCESS;
  payload: {ImgurResult: IImgurResult};
}

export interface IGetGoogleResult {
  type: Types.MASTER_GET_GOOGLE_RESULT;
  payload: {query: string, lang: string};
}

export interface IGetGoogleImgResult {
  type: Types.MASTER_GET_GOOGLE_IMG_RESULT;
  payload: {imglink: string, lang: string};
}

export interface IGetGoogleImgResultSuccess {
  type: Types.MASTER_GET_GOOGLE_IMG_RESULT_SUCCESS;
  payload: {GoogleImgResult: IGoogleImgResult};
}

export interface IGetGoogleResultSuccess {
  type: Types.MASTER_GET_GOOGLE_RESULT_SUCCESS;
  payload: {GoogleResult: IGoogleResult};
}

export interface IGetInformationSuccess {
  type: Types.MASTER_GET_INFORMATION_RESULT_SUCCESS;
  payload: {Information: ISearchInformation};
}

export interface IGetInformationButtonsSuccess {
  type: Types.MASTER_GET_INFORMATION_BUTTONS_SUCCESS;
  payload: ILocationButtons[]
}

export interface IGetTopSightsSuccess {
  type: Types.MASTER_GET_TOPSIGHTS_SUCCESS;
  payload: {TopSights: ITopSights}
}


export interface IGetCityLatLong {
  type: Types.MASTER_GET_CITY_LATLONG;
  payload: {cityname: string};
}

export interface IGetCityLatLongSuccess {
  type: Types.MASTER_GET_CITY_LATLONG_SUCCESS;
  payload: ICityLatLong[]
}

export interface IAppStart {
  type: Types.MASTER_APP_START;
  payload: {};
}

export interface IFetchingCustomer {
  type: Types.MASTER_FETCHING_CUSTOMERS;
}

export interface IFetchingTmpCustomer {
  type: Types.MASTER_FETCHING_TMP_CUSTOMERS;
  payload: {unitId: string};
}

export interface IFetchCustomerSuccess {
  type: Types.MASTER_FETCH_CUSTOMERS_SUCCESS;
  payload: ICustomerInfo[];
}

export interface IFetchCustomerTmpSuccess {
  type: Types.MASTER_FETCH_CUSTOMERS_TMP_SUCCESS;
  payload: ICustomerTmpInfo[];
}

export interface IUpdateCustomerDropdownData {
  type: Types.MASTER_UPDATE_CUSTOMER_DROPDOWN_DATA;
  payload: DropdownItemType[];
}

export interface IUpdateCustomer {
  type: Types.MASTER_UPDATE_CUSTOMER;
  payload: ICustomerDetail | null;
}

export interface ICreateCustomer {
  type: Types.MASTER_CREATE_CUSTOMER;
  payload: {model: ICustomerDetail};
}

export interface IUpdateFarmLocation {
  type: Types.MASTER_UPDATE_FARM_LOCATION;
  payload: {farmcode: string, region: string}
}

export interface IFetchCheckingHistory {
  type: Types.MASTER_FETCH_CHECKING_HISTORY;
  payload: ICheckingHistoryModel;
}

export interface IFetchCheckingHistorySucess {
  type: Types.MASTER_FETCH_CHECKING_HISTORY_SUCCESS;
  payload: ICheckingHistoryItem[];
}

export interface IFetchAllReport {
  type: Types.MASTER_FETCH_ALL_REPORT;
  payload: IAllReportModal;
}

export interface IFetchAllReportSuccess {
  type: Types.MASTER_FETCH_ALL_REPORT_SUCCESS;
  payload: IAllReport[]
}

export interface IFetchHistoryDetail {
  type: Types.MASTER_FETCH_HISTORY_DETAIL;
  payload: ICheckingHistoryDetail[];
}

export interface IFetchHistoryDetailSuccess {
  type: Types.MASTER_FETCH_HISTORY_DETAIL_SUCCESS;
  payload: ICheckingHistoryDetail[];
}

export interface IFetchingFarmLocationList {
  type: Types.MASTER_FETCHING_FARMLIST;
  payload: {regionId: string};
}

export interface IGetCheckingImage {
  type: Types.MASTER_GET_CHECKING_IMAGE;
  payload: {imageID: string, isCheckOut: boolean};
}

export interface IGetCheckingImageSuccess {
  type: Types.MASTER_GET_CHECKING_IMAGE_SUCCESS;
  payload: {checkingImage: string, isCheckOut: boolean, imageID: string}
}

export interface IFetchingFarmLocationListSuccess {
  type: Types.MASTER_FETCHING_FARMLIST_SUCCESS;
  payload: {farmlist : IFarmItem[]};
}

export interface IUserLocLatLong {
  type: Types.MASTER_GET_USER_LATLONG;
  payload: {location : IUserLatLong};
}

export interface IUpdateUserLocLatLong {
  type: Types.MASTER_UPDATE_USER_LATLONG;
  payload: IUserLatLongModel;
}

export interface IGetLastCheckinID {
  type: Types.MASTER_GET_LAST_CHECKID;
  payload: {email: string, farmcode: string}
}

export interface IGetLastCheckinIDSuccess {
  type: Types.MASTER_GET_LAST_CHECKID_SUCCESS;
  payload: ILastCheckinResponse[];
}
// ...
export interface IGenerateQRSuccess {
  type: Types.MASTER_GENERATE_QR_SUSCCESS;
  payload: {FarmID: string,QRcode: string};
}

export interface IUpdateCheckingDay {
  type: Types.MASTER_UPDATE_CHECKING_DAY;
  payload: {checkingId : number, Date_Qty: number };
}

export interface IUpdateCheckingNote {
  type: Types.MASTER_UPDATE_CHECKING_NOTE;
  payload: {checkingId: number, Note: string};
}

export interface IUploadImageToApi {
  type: Types.MASTER_UPLOAD_IMAGE_TO_API;
  payload: IImageModel;
}

export interface ICreateNewCheckingSuccess {
  type: Types.MASTER_CREATE_CHECKING_SUCCESS;
  payload: {checkingId: number};
}

export interface ICreateNewCheckoutSuccess {
  type: Types.MASTER_CREATE_CHECKOUT_SUCCESS;
  payload: {checkoutId: number};
}

export interface ICreateCustomerSuccess {
  type: Types.MASTER_CREATE_CUSTOMER_SUCCESS;
  payload: {model: ICustomerTmpInfo};
}

export interface IPostCustomerToEsigning {
  type: Types.MASTER_POST_CUSTOMER_TO_ESIGNING;
  payload: ICustomerDetail;
}

export interface IUploadImageToOneDrive {
  type: Types.MASTER_UPLOAD_IMAGE_TO_ONEDRIVE;
  payload: IImageChecking;
}

export interface IUploadImageToOneDriveSuccess {
  type: Types.MASTER_UPLOAD_IMAGE_TO_ONEDRIVE_SUCCESS;
  payload: {response: IImageResponse};
}

//

export interface IMasterChangeSubmitStatus {
  type: Types.MASTER_CHANGE_IS_SUBMIT_STATUS;
  payload: {isResult: boolean; screen: string | undefined};
}


export interface IFetchingLocationByUnits {
  type: Types.MASTER_FETCHING_LOCAL_BY_UNITIDS;
  payload: string[];
}

export interface IUpdateLocationDropdownData {
  type: Types.MASTER_UPDATE_LOCATION_DROPDOWN_DATA;
  payload: DropdownItemType[];
}

export interface IFetchLocationByUnitsSuccess {
  type: Types.MASTER_FETCH_LOCATION_BY_UNIT_SUCCESS;
  payload: ILocation[];
}

export interface IFetchingOffice {
  type: Types.MASTER_FETCHING_OFFICE_BY_USER;
  payload: string;
}

export interface IFetchOfficeSuccess {
  type: Types.MASTER_FETCH_OFFICE_BY_USER_SUCCESS;
  payload: IMasterResponseCommon[];
}

export interface IFetchingDepartment {
  type: Types.MASTER_FETCHING_DEPARTMENT_BY_USER;
  payload: IDepartmentParams;
}

export interface IFetchDepartmentSuccess {
  type: Types.MASTER_FETCH_DEPARTMENT_BY_USER_SUCCESS;
  payload: IMasterResponseCommon[];
}

export interface IFetchingProductByUnit {
  type: Types.MASTER_FETCHING_PRODUCTS_BY_UNIT;
  payload?: string;
}

export interface IFetchProductsByUnitSuccess {
  type: Types.MASTER_FETCH_PRODUCTS_BY_UNIT_SUCCESS;
  payload: IMasterResponseCommon[];
}

export interface IGetSwinePrice {
  type: Types.MASTER_GET_SWINE_PRICE_BY_CUST_PRODUCT;
  payload: IPriceModel;
}

export interface IGetSwinePriceSuccess {
  type: Types.MASTER_GET_SWINE_PRICE_BY_CUST_PRODUCT_SUCCESS;
  payload: IPrice[];
}

export interface IGetCustomerBalance {
  type: Types.MASTER_GET_CUSTOMER_BALANCES;
  payload: {custId: string; unitId: string};
}

export interface IGetCustomerBalanceSuccess {
  type: Types.MASTER_GET_CUSTOMER_BALANCES_SUCCESS;
  payload: ICustomerBalance[];
}

export interface IGetUnitOfCustomer {
  type: Types.MASTER_GET_UNITS_OF_CUSTOMER;
  payload: {custId: string; units: string[]};
}

export interface IGetUnitOfCustomerSuccess {
  type: Types.MASTER_GET_UNITS_OF_CUSTOMER_SUCCESS;
  payload: {units: IUnit[]; unitsDropdownData: DropdownItemType[]};
}

export interface IGetLocationByUser {
  type: Types.MASTER_GET_LOCATIONS_BY_USER;
  payload: {};
}

export interface IGetLocationByUserSuccess {
  type: Types.MASTER_GET_LOCATIONS_BY_USER_SUCCESS;
  payload: {locations: ILocation[]};
}

export interface IGetMenuByUser {
  type: Types.MASTER_GET_MENU_BY_USER;
  payload: {};
}

export interface IGetMenuByUserWithoutPass {
  type: Types.MASTER_GET_MENU_BY_USER_WITHOUT_PASS;
  payload : {email : string};
}

export interface IGetMenuByUserSuccess {
  type: Types.MASTER_GET_MENU_BY_USER_SUCCESS;
  payload: {menus: IMasterMenu[]; userMenuIds: number[]};
}
// ...
export interface IGetSaleLocationSuccess {
  type: Types.MASTER_GET_SALE_LOCATION_SUCCESS;
  payload: {locations: ILocation[]};
}

export interface IGetSaleUnits {
  type: Types.MASTER_GET_SALE_UNITS;
  payload: {custId?: string};
}

export interface IGetSaleUnitsSuccess {
  type: Types.MASTER_GET_SALE_UNITS_SUCCESS;
  payload: {units: IUnit[]};
}

export interface IGetAllProducts {
  type: Types.MASTER_GET_ALL_PRODUCTS;
  payload: {};
}

export interface IGetAllProductsSuccess {
  type: Types.MASTER_GET_ALL_PRODUCTS_SUCCESS;
  payload: {products: IProductStockDTO[]};
}

export interface ISendMessagesToChatGPT {
  type: Types.MASTER_SEND_MESSAGES_TO_CHATGPT;
  payload: {prompt: string, max_tokens: number};
}

export interface ISendMessagesToChatGPTSuccess {
  type: Types.MASTER_SEND_MESSAGES_TO_CHATGPT_SUCCESS;
  payload: IChatGPTResponse[]
}

export type MasterActionsType =
  | ISendMessagesToChatGPTSuccess
  | IGetImgUrlSuccess
  | IGetGoogleResultSuccess
  | IGetGoogleImgResultSuccess
  | IGetInformationSuccess
  | IGetTopSightsSuccess
  | IGetInformationButtonsSuccess
  | IGetCityLatLongSuccess
  | IFetchCustomerSuccess
  | IFetchCustomerTmpSuccess
  | IFetchingFarmLocationListSuccess
  | IGetLastCheckinIDSuccess
  | IGetCheckingImageSuccess
  | IFetchHistoryDetailSuccess
  | IFetchCheckingHistorySucess
  | IFetchAllReportSuccess
  | IUploadImageToOneDriveSuccess
  | IMasterChangeSubmitStatus
  | IFetchingLocationByUnits
  | IFetchLocationByUnitsSuccess
  | IFetchOfficeSuccess
  | IFetchDepartmentSuccess
  | IFetchProductsByUnitSuccess
  | IGetSwinePriceSuccess
  | IGetCustomerBalanceSuccess
  | IUpdateCustomerDropdownData
  | IUpdateCustomer
  | ICreateCustomerSuccess
  | ICreateNewCheckingSuccess
  | IGenerateQRSuccess
  | ICreateNewCheckoutSuccess
  | IUpdateLocationDropdownData
  | IGetUnitOfCustomerSuccess
  | IGetLocationByUserSuccess
  | IGetMenuByUserSuccess
  | IGetSaleLocationSuccess
  | IGetSaleUnitsSuccess
  | IUserLocLatLong
  | IGetAllProductsSuccess;